fixing some map leaks

+++ libft/ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: jbelkerf <jbelkerf@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/21 16:07:18 by jbelkerf          #+#    #+#             */
-/*   Updated: 2024/11/04 16:40:57 by jbelkerf         ###   ########.fr       */
+/*   Updated: 2025/02/26 16:23:37 by jbelkerf         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,10 +18,10 @@
  * 1- if size less then len_d  ten the len_d == size return is len_s + size
  * 2- if in size there's a room for src or a part of it then we return default
  */
-size_t	ft_strlcat(char *dst, const char *src, size_t size)
+int	ft_strlcat(char *dst, const char *src, int size)
 {
-	size_t	i;
-	size_t	j;
+	int	i;
+	int	j;
 
 	if (size == 0)
 		return (ft_strlen(src));
